requirements.txt
click==7.1.2
config==0.5.0.post0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
pkg-resources==0.0.0
python-dotenv==0.14.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
